<test>: 그룹 멤버 추방 테스트 추가 (응답 코드 오류 수정)

+++ src/main/java/org/unisphere/unisphere/group/service/GroupCommandService.java
@@ -286,8 +286,11 @@ public class GroupCommandService {
 			throw ExceptionStatus.NOT_GROUP_ADMIN.toServiceException();
 		}
 		GroupRegistration targetRegistration = groupRegistrationRepository
-				.findByGroupIdAndMemberIdAndRegisteredAtNull(group.getId(), targetMemberId)
+				.findByGroupIdAndMemberId(group.getId(), targetMemberId)
 				.orElseThrow(ExceptionStatus.NOT_GROUP_REQUEST_MEMBER::toServiceException);
+		if (targetRegistration.getRegisteredAt() != null) {
+			throw ExceptionStatus.ALREADY_APPROVED_MEMBER.toServiceException();
+		}
 		groupRegistrationRepository.delete(targetRegistration);
 	}
 }
